+++ app/Models/Results/EvaluationStrategies/EvaluationVyfuk2011.php
@@ -24,7 +24,7 @@ class EvaluationVyfuk2011 extends EvaluationStrategy
         return 's.raw_points';
     }
 
-    public function getTaskPoints(TaskModel $task, ModelCategory $category): int
+    public function getTaskPoints(TaskModel $task, ModelCategory $category): float
     {
         return $task->points;
     }
